Disable input duplicate name
Currently there's no checking on input name. Thus, duplicate name input is
possible. In a draw from hat it is good to disable duplication to avoid
confusion.

Add `nameSet` variable to contain set of inputted name. Set is used because it's
API is easy to use and suitable for this use case. Moreover, it is [already
supported by all browser](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set).

The set will be inserted on name insert iteration, only if the name isn't
available in the set. The set will be removed on `clearList` event.